requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.10

# Redis
redis[hiredis]==5.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17

# AWS SDK
boto3==1.35.36
aioboto3==13.2.0

# OpenAI
openai==1.54.3

# ElevenLabs TTS
elevenlabs==1.9.0

# PDF Generation
reportlab==4.2.4
Pillow==10.4.0

# Celery for async tasks
celery[redis]==5.4.0
flower==2.0.1

# HTTP Client
httpx==0.27.2
aiohttp==3.10.10

# Utilities
python-dotenv==1.0.1
email-validator==2.2.0

# Monitoring & Logging
sentry-sdk[fastapi]==2.17.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.10.0
ruff==0.7.3
